training/requirements.txt
tensorflow>=2.15
numpy
Pillow
matplotlib
scikit-learn
kaggle
python-dotenv
